+++ storyblok/ImageTextSection.vue
@@ -21,26 +21,9 @@ const blurredImage = computed(() => getOptimizedImage(props.blok.image, 1000, 0,
     <div
       class="container grid items-center gap-6 sm:gap-10 md:gap-12 lg:grid-cols-2"
     >
-      <div
-        class="order-last text-left"
-        :class="blok.reverse_layout ? '' : 'lg:order-first'"
-      >
-        <Eyebrow v-if="blok.eyebrow">{{ blok.eyebrow }}</Eyebrow>
-        <Headline v-if="blok.headline" size="small" :headline="blok.headline" />
-        <div v-if="blok.text" class="prose prose-lg mb-6">
-          <StoryblokRichText :doc="blok.text" />
-        </div>
-        <div v-if="blok?.buttons?.length" class="flex flex-col items-start justify-start gap-4 sm:flex-row">
-          <Button
-            v-for="button in blok.buttons"
-            :key="button._uid"
-            :button="button"
-          />
-        </div>
-      </div>
       <div>
         <div v-if="blok.image.filename" class="relative rounded-xl bg-cover bg-center bg-no-repeat p-8 md:p-16" :style="`background: url('${blurredImage}'); background-size: cover; background-repeat: no-repeat; background-position: center;`">
-          <DecorationImageTopLeft class="absolute left-0 top-0 translate-x-[-5px] translate-y-[-5px] scale-50 md:translate-x-[25px] md:translate-y-[20px] md:scale-100" />
+          <DecorationImageTopLeft class="absolute left-0 top-0 origin-top-left translate-x-[-20px] translate-y-[-25px] scale-50 md:translate-x-[25px] md:translate-y-[20px] md:scale-100" />
           <img
             v-if="optimizedImages"
             :src="optimizedImages.mobile"
@@ -58,6 +41,23 @@ const blurredImage = computed(() => getOptimizedImage(props.blok.image, 1000, 0,
             :src="optimizedImages.desktop"
             :alt="blok.image.alt"
             class="invisible hidden rounded-lg lg:visible lg:block"
+          />
+        </div>
+      </div>
+      <div
+        class="text-left"
+        :class="[blok.reverse_mobile_layout ? 'order-first' : '', blok.reverse_desktop_layout ? 'lg:order-last' : 'lg:order-first']"
+      >
+        <Eyebrow v-if="blok.eyebrow">{{ blok.eyebrow }}</Eyebrow>
+        <Headline v-if="blok.headline" size="small" :headline="blok.headline" />
+        <div v-if="blok.text" class="prose prose-lg mb-6">
+          <StoryblokRichText :doc="blok.text" />
+        </div>
+        <div v-if="blok?.buttons?.length" class="flex flex-col items-start justify-start gap-4 sm:flex-row">
+          <Button
+            v-for="button in blok.buttons"
+            :key="button._uid"
+            :button="button"
           />
         </div>
       </div>